pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "st-assignment"
version = "0.1.0"
description = "Network Health Platform — PoC"
readme = {text = "Assignment project", content-type = "text/plain"}
requires-python = ">=3.11, <3.12"
dependencies = []

[project.optional-dependencies]
core = [
    "numpy<2",
    "kafka-python==2.0.2",
    "pyspark==3.5.8",
    "pyiceberg[s3]==0.7.1",
    "clickhouse-connect==0.7.19",
    "pydantic-settings==2.3.4",
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "prometheus-client==0.20.0",
    "duckdb==1.0.0",
    "scikit-learn==1.5.0",
    "joblib==1.4.2",
    "pandas==2.2.2",
    "cryptography==46.0.7",
    "pyarrow==14.0.0",
    "apache-airflow==2.9.1",
]

airflow = [
    "kafka-python==2.0.2",
    "pyiceberg[s3]==0.7.1",
    "pyarrow==14.0.0",
]

iceberg-init = [
    "pyiceberg[s3]==0.7.1",
    "pyspark==3.5.8",
]


[dependency-groups]
dev = [
    "pytest==8.2.2",
    "pytest-cov==5.0.0",
    "ruff==0.4.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
